fix bug that allowed overlapping disks
John Shier and Chris Ennis reported that the Hurwitz program showed
overlapping disks when s and a were both assigned large values. The
problem was caused by two errors. First, 'bigLimit' was being set to
'1 / cellRowsAndCols', whereas it should be 'boxSide/cellRowsAndCols'.
This had the effect of suppressing overlaps when boxSide > 1 (the
usual case), although it also reduced efficiency. The second bug is
that bigLimit should actually be boxSide/cellRowsAndCols/2 -- we want
the disks in a database cell to have a diameter no greater than the
width of the cell.

In addition to fixing the bug, added code to visualize the cells
and to verify that no overlaps exist.

This commit includes console.log messages to be removed before merge.