Batched searching to avoid frozen UI
The program will make up to 10^7 attempts to find an available spot for
a new disk. That can take several seconds, and during that period the UI
becomes unresponsive (heartbeat on Running message becomes jerky, buttons
do not respond to clicks). This commit divides the task into 1,000 batches
of 10,000 attempts each, allowing other code in the event queue to run.