New program 9a for Hurwitz zeta
John Shier requested a version of Program 9 (one disk at a time,
showing the halo or buffer of excluded area around each disk) using
the Hurwitz rather than the Riemann zeta function. This is for the
GitHub-pages version of the program; it will not be ported to master,
which is tracked by the bit-player.org posting.